Creates a navbar at the bottom that cycles between two fragments: one for all the articles loaded and another for articles saved to read later.
