+++ app/src/main/res/layout/activity_main.xml
@@ -6,21 +6,19 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <LinearLayout
+    <FrameLayout
+        android:id="@+id/fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@color/colorPrimary"
-        android:orientation="horizontal"
-        tools:layout_editor_absoluteX="46dp"
-        tools:layout_editor_absoluteY="217dp">
+        android:layout_above="@+id/bottom_navigation"/>
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/article_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:padding="4dp"
-            android:scrollbars="vertical" />
 
-    </LinearLayout>
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/bottom_navigation"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:background="?android:attr/windowBackground"
+        app:menu="@menu/bottom_navigation"/>
 
 </RelativeLayout>
