+++ app/src/main/res/layout/fragment_readlater.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent" android:layout_height="match_parent"
+    android:background="@color/colorPrimary">
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/readlater_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:padding="4dp"
+        android:scrollbars="vertical" />
+    <TextView
+        android:id="@+id/no_articles"
+        android:layout_width="match_parent"
+        android:layout_height="300dp"
+        android:textSize="30dp"
+        android:textColor="@color/colorFont"
+        android:text="@string/no_articles_saved"
+        />
+</RelativeLayout>
