+++ app/src/main/res/menu/bottom_navigation.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_home"
+        android:icon="@drawable/ic_home_black_24dp"
+        android:title="@string/home" />
+    <item
+        android:id="@+id/nav_read_later"
+        android:icon="@drawable/ic_save_white"
+        android:title="@string/read_later"/>
+
+</menu>
